exercise06: fix contact bounce of user_button with timer interrupt
Timer TIM10's counter is set to 0 on key_press in main routine and
in 50 ms happens its update interrupt. The assosiated isr checks if
key is still pressed and sets user_key_down flag.
If key is not pressed, main routine checks for user_key_down flag and
if it is set, changes led state.

Signed-off-by: Yaroslav Fedoriachenko <yar.fed99@gmail.com>